+++ src/Sentry.Unity/TimeDebounceBase.cs
@@ -2,6 +2,10 @@ using System;
 
 namespace Sentry.Unity;
 
+/// <summary>
+/// Interface for debouncing Unity log messages.
+/// </summary>
+[Obsolete("Use IErrorEventThrottler and ContentBasedThrottler instead. This interface will be removed in a future version.")]
 public interface IUnityLogMessageDebounce
 {
     bool Debounced();
@@ -10,6 +14,7 @@ public interface IUnityLogMessageDebounce
 /// <summary>
 /// This class is not thread-safe and is designed to be called by Unity non-threaded logger callback
 /// </summary>
+[Obsolete("Use ContentBasedThrottler instead. This class will be removed in a future version.")]
 internal class TimeDebounceBase : IUnityLogMessageDebounce
 {
     private static DateTimeOffset Now => DateTimeOffset.UtcNow;
@@ -33,6 +38,7 @@ internal class TimeDebounceBase : IUnityLogMessageDebounce
 /// <summary>
 /// This class is not thread-safe and is designed to be called by Unity non-threaded logger callback
 /// </summary>
+[Obsolete("Use ContentBasedThrottler instead. This class will be removed in a future version.")]
 internal sealed class LogTimeDebounce : TimeDebounceBase
 {
     public LogTimeDebounce(TimeSpan debounceOffset) => DebounceOffset = debounceOffset;
@@ -41,6 +47,7 @@ internal sealed class LogTimeDebounce : TimeDebounceBase
 /// <summary>
 /// This class is not thread-safe and is designed to be called by Unity non-threaded logger callback
 /// </summary>
+[Obsolete("Use ContentBasedThrottler instead. This class will be removed in a future version.")]
 internal sealed class ErrorTimeDebounce : TimeDebounceBase
 {
     public ErrorTimeDebounce(TimeSpan debounceOffset) => DebounceOffset = debounceOffset;
@@ -49,6 +56,7 @@ internal sealed class ErrorTimeDebounce : TimeDebounceBase
 /// <summary>
 /// This class is not thread-safe and is designed to be called by Unity non-threaded logger callback
 /// </summary>
+[Obsolete("Use ContentBasedThrottler instead. This class will be removed in a future version.")]
 internal sealed class WarningTimeDebounce : TimeDebounceBase
 {
     public WarningTimeDebounce(TimeSpan debounceOffset) => DebounceOffset = debounceOffset;