+++ test/Sentry.Unity.Editor.Tests/ScriptableSentryUnityOptionsTests.cs
@@ -21,7 +21,8 @@ public class ScriptableSentryUnityOptionsTests
         StringAssert.Contains("Enabled", optionsAsString);
         StringAssert.Contains("Dsn", optionsAsString);
         StringAssert.Contains("CaptureInEditor", optionsAsString);
-        StringAssert.Contains("EnableLogDebouncing", optionsAsString);
+        StringAssert.Contains("EnableErrorEventThrottling", optionsAsString);
+        StringAssert.Contains("ThrottleDedupeWindow", optionsAsString);
         StringAssert.Contains("TracesSampleRate", optionsAsString);
         StringAssert.Contains("AutoSessionTracking", optionsAsString);
         StringAssert.Contains("AutoSessionTrackingInterval", optionsAsString);